# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.Code:
    enum:
    - 0
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 2001
    - 2002
    - 2003
    - 2004
    - 3001
    - 3002
    - 3003
    - 3004
    - 4001
    - 4002
    - 4003
    - 4004
    type: integer
    x-enum-varnames:
    - Success
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrInvalidParams
    - ErrBindJSON
    - ErrBindQuery
    - ErrBindForm
    - ErrDatabase
    - ErrRecordNotFound
    - ErrRecordExists
    - ErrRecordInvalidate
    - ErrTokenInvalid
    - ErrTokenExpired
    - ErrTokenMissing
    - ErrPermissionDenied
...
  model.AuditLog:
    properties:
      action:
        description: 操作动作（如：create, update, delete, login）
        type: string
      created_at:
        type: string
      duration:
        description: 请求耗时（毫秒）
        type: integer
      error:
        description: 错误信息
        type: string
      extra:
        description: 额外信息（JSON）
        type: string
      id:
        type: integer
      ip:
        description: 客户端 IP
        type: string
      method:
        description: HTTP 方法
        type: string
      path:
        description: 请求路径
        type: string
      request:
        description: 请求体（可选）
        type: string
      resource:
        description: 操作资源（如：user, file, role）
        type: string
      resource_id:
        description: 资源ID
        type: string
      response:
        description: 响应体（可选）
        type: string
      status_code:
        description: HTTP 状态码
        type: integer
      updated_at:
        type: string
      user_agent:
        description: User Agent
        type: string
      user_id:
        description: 用户ID
        type: integer
      username:
        description: 用户名
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        $ref: '#/definitions/errors.Code'
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cccvno1/nova
  description: |-
    Nova 是一个基于 Go + Echo 的现代化 Web 应用框架，提供用户认证、RBAC 权限管理、文件上传、异步任务队列、审计日志等功能。

    主要特性：
    - 用户认证与授权（JWT）
    - RBAC 权限管理（基于 Casbin）
    - 文件上传与管理（支持本地/OSS/S3）
    - 异步任务队列（基于 Redis）
    - 定时任务调度（基于 Cron）
    - 审计日志（自动记录所有操作）
    - 限流保护（IP/用户维度）
    - 健康检查与监控
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Nova API
  version: "1.0"
paths:
  /audit-logs:
    get:
      consumes:
      - application/json
      description: 查询审计日志列表，支持多条件过滤和分页
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 操作类型
        in: query
        name: action
        type: string
      - description: 资源类型
        in: query
        name: resource
        type: string
      - description: IP地址
        in: query
        name: ip
        type: string
      - description: HTTP方法
        in: query
        name: method
        type: string
      - description: 请求路径
        in: query
        name: path
        type: string
      - description: HTTP状态码
        in: query
        name: status_code
        type: integer
      - description: 开始时间(RFC3339格式)
        in: query
        name: start_time
        type: string
      - description: 结束时间(RFC3339格式)
        in: query
        name: end_time
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 审计日志列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AuditLog'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取审计日志列表
      tags:
      - 审计日志
  /audit-logs/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取审计日志的详细信息
      parameters:
      - description: 审计日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 审计日志详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AuditLog'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 审计日志不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取审计日志详情
      tags:
      - 审计日志
  /audit-logs/clean:
    delete:
      consumes:
      - application/json
      description: 清理指定天数之前的审计日志（物理删除）
      parameters:
      - default: 90
        description: 保留天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除结果(包含删除数量和截止时间)
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 清理旧审计日志
      tags:
      - 审计日志
  /audit-logs/stats:
    get:
      consumes:
      - application/json
      description: 获取审计日志的综合统计信息，包括总数、成功/失败数、操作统计、用户统计和资源统计
      parameters:
      - default: 7天前
        description: 开始时间(RFC3339格式)
        in: query
        name: start_time
        type: string
      - default: 当前时间
        description: 结束时间(RFC3339格式)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 统计信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取综合统计
      tags:
      - 审计日志
  /audit-logs/stats/actions:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的操作类型统计
      parameters:
      - default: 24小时前
        description: 开始时间(RFC3339格式)
        in: query
        name: start_time
        type: string
      - default: 当前时间
        description: 结束时间(RFC3339格式)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作统计列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取操作统计
      tags:
      - 审计日志
  /audit-logs/stats/resources:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的资源操作统计
      parameters:
      - default: 7天前
        description: 开始时间(RFC3339格式)
        in: query
        name: start_time
        type: string
      - default: 当前时间
        description: 结束时间(RFC3339格式)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 资源统计列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取资源统计
      tags:
      - 审计日志
  /audit-logs/stats/users:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内操作最频繁的用户统计（Top N）
      parameters:
      - default: 7天前
        description: 开始时间(RFC3339格式)
        in: query
        name: start_time
        type: string
      - default: 当前时间
        description: 结束时间(RFC3339格式)
        in: query
        name: end_time
        type: string
      - default: 10
        description: 返回数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户统计列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户统计
      tags:
      - 审计日志
  /audit-logs/user/{userId}:
    get:
      consumes:
      - application/json
      description: 查询指定用户的审计日志列表
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 审计日志列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AuditLog'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户审计日志
      tags:
      - 审计日志
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"